+++ packages/influxdb/src/lib/types/AnalyticsSchema.ts
@@ -1,0 +1,32 @@
+export enum Points {
+	Guilds = 'guilds',
+	Users = 'users',
+	Commands = 'commands',
+	Memory = 'memory',
+	MessageCount = 'message_count'
+}
+
+export enum Tags {
+	Shard = 'shard',
+	Client = 'client_id',
+	Action = 'action',
+	OriginEvent = 'origin_event'
+}
+
+export enum CommandRunTypes {
+	Message = 'message',
+	ChatInput = 'chat-input',
+	ContextMenu = 'context-menu',
+	CommandAutocomplete = 'command-autocomplete'
+}
+
+export enum CommandTypes {
+	Category = 'category',
+	RunType = 'run_type'
+}
+
+export enum Actions {
+	Addition = 'addition',
+	Subtraction = 'subtraction',
+	Sync = 'sync'
+}