+++ packages/subcommands-advanced/src/index.ts
@@ -1,3 +1,5 @@
+import 'tslib';
+
 export { subCommandsRegistry, subCommandsGroupRegistry } from './lib/utils/functions';
 export * from './lib/structures/command';
 export * from './lib/utils/const';