+++ packages/env/CHANGELOG.md
@@ -3,6 +3,40 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.9](https://github.com/kaname-png/neko-plugins/compare/@kaname-png/plugin-env@1.1.7...@kaname-png/plugin-env@1.1.9) (2023-10-22)
+
+**Note:** Version bump only for package @kaname-png/plugin-env
+
+## [1.1.8](https://github.com/kaname-png/neko-plugins/compare/@kaname-png/plugin-env@1.1.7...@kaname-png/plugin-env@1.1.8) (2023-09-06)
+
+**Note:** Version bump only for package @kaname-png/plugin-env
+
+## [1.1.7](https://github.com/kaname-png/neko-plugins/compare/@kaname-png/plugin-env@1.1.6...@kaname-png/plugin-env@1.1.7) (2023-06-08)
+
+### Bug Fixes
+
+-   tests ([589179f](https://github.com/kaname-png/neko-plugins/commit/589179f2021a4cd6054a7ee064e4e40a26a7ba94))
+
+## [1.1.6](https://github.com/kaname-png/neko-plugins/compare/@kaname-png/plugin-env@1.1.5...@kaname-png/plugin-env@1.1.6) (2023-06-05)
+
+### Bug Fixes
+
+-   tslib import ([ce34a4d](https://github.com/kaname-png/neko-plugins/commit/ce34a4da81c147528bb128e3681f1d5039c134ba))
+
+## [1.1.5](https://github.com/kaname-png/neko-plugins/compare/@kaname-png/plugin-env@1.1.4...@kaname-png/plugin-env@1.1.5) (2022-12-29)
+
+**Note:** Version bump only for package @kaname-png/plugin-env
+
+## [1.1.4](https://github.com/kaname-png/neko-plugins/compare/@kaname-png/plugin-env@1.1.3...@kaname-png/plugin-env@1.1.4) (2022-12-29)
+
+### Bug Fixes
+
+-   import typing resolution ([41f47ff](https://github.com/kaname-png/neko-plugins/commit/41f47ffc58d8b8ebe4a06804ed736eda7f19f12a))
+
+## [1.1.3](https://github.com/kaname-png/neko-plugins/compare/@kaname-png/plugin-env@1.1.2...@kaname-png/plugin-env@1.1.3) (2022-12-04)
+
+**Note:** Version bump only for package @kaname-png/plugin-env
+
 ## [1.1.2](https://github.com/kaname-png/neko-plugins/compare/@kaname-png/plugin-env@1.1.1...@kaname-png/plugin-env@1.1.2) (2022-05-06)
 
 **Note:** Version bump only for package @kaname-png/plugin-env